+++ packages/transaction-pay-controller/src/strategy/across/across-quotes.ts
@@ -193,7 +193,7 @@ async function normalizeQuote(
   request: QuoteRequest,
   fullRequest: PayStrategyGetQuotesRequest,
 ): Promise<TransactionPayQuote<AcrossQuote>> {
-  const { accountSupports7702, messenger } = fullRequest;
+  const { messenger } = fullRequest;
   const { quote } = original;
 
   const { usdToFiatRate, sourceFiatRate, targetFiatRate } = getFiatRates(
@@ -208,7 +208,6 @@ async function normalizeQuote(
     quote,
     messenger,
     request,
-    accountSupports7702,
   );
 
   const targetNetwork = getFiatValueFromUsd(new BigNumber(0), usdToFiatRate);
@@ -391,7 +390,6 @@ async function calculateSourceNetworkCost(
   quote: AcrossSwapApprovalResponse,
   messenger: TransactionPayControllerMessenger,
   request: QuoteRequest,
-  accountSupports7702: boolean,
 ): Promise<{
   sourceNetwork: TransactionPayQuote<AcrossQuote>['fees']['sourceNetwork'];
   gasLimits: AcrossGasLimits;
@@ -414,7 +412,6 @@ async function calculateSourceNetworkCost(
       to: transaction.to,
       value: transaction.value ?? '0x0',
     })),
-    accountSupports7702,
   });
   const { batchGasLimit, is7702 } = gasEstimates;